+++ Samples~/Demo/Code/Scripts/TimerExample.cs
@@ -12,13 +12,18 @@ namespace PsigenVision.ImprovedTimers.Demo
         private CountdownTimer timer1, timer2;
         [SerializeField] private float outerRadialBarProgress;
         [SerializeField] private float innerRadialBarProgress;
-
+        private int outerRadialBarSegmentCount;
         private bool isInitialized = false;
+        void Awake() => Initialize();
         
         // Start is called once before the first execution of Update after the MonoBehaviour is created
         void Start()
         {
-            Initialize();
+            if (!isInitialized) return;
+            
+            //Start countdown timers
+            timer1.Start();
+            timer2.Start();
         }
 
         private void Initialize()
@@ -59,20 +64,18 @@ namespace PsigenVision.ImprovedTimers.Demo
                 Debug.LogError("Failed to create timer2");
                 isInitialized = false;
             }
-            
-            //Start countdown timers
-            timer1.Start();
-            timer2.Start();
-            
-            //Provide debug log message hooks into timer start and stop callback events
-            timer1.OnTimerStart += () => Debug.Log("Timer1 started");
-            timer2.OnTimerStart += () => Debug.Log("Timer2 started");
-            timer1.OnTimerStop += () => Debug.Log("Timer1 stopped");
-            timer2.OnTimerStop += () => Debug.Log("Timer2 stopped");
 
-            if (!isInitialized)
+            if (!isInitialized) Debug.LogError("Failed to initialize timer example");
+            else
             {
-                Debug.LogError("Failed to initialize timer example");
+                //Calculate the amount of segments in outer radial bar
+                outerRadialBarSegmentCount = Mathf.CeilToInt(timer1Duration); //Derive how many segments we want in the bar
+                
+                //Provide debug log message hooks into timer start and finish callback events
+                timer1.OnTimerStart += () => Debug.Log("Timer1 started");
+                timer2.OnTimerStart += () => Debug.Log("Timer2 started");
+                timer1.OnTimerFinish += () => Debug.Log("Timer1 finished");
+                timer2.OnTimerFinish += () => Debug.Log("Timer2 finished");
             }
         }
 
@@ -82,9 +85,11 @@ namespace PsigenVision.ImprovedTimers.Demo
             if (!isInitialized) return;
             //Update radial bars' progress with the countdown timers' progress
             //First radial bar will be treated as a segmented radial bar
-            float maxSegments = Mathf.CeilToInt(timer1Duration); //Derive how many segments we want in the bar
-            float segmentsToShow = Mathf.CeilToInt(timer1.Progress * maxSegments); //Derive how many segments are currently visible based on the timer's progress
-            outerRadialBarProgress = (segmentsToShow / maxSegments)*100f;
+            //Derive how many segments are currently visible based on the timer's progress by dividing by the total segment count
+            outerRadialBarProgress = 
+                (float)Mathf.CeilToInt(timer1.Progress * outerRadialBarSegmentCount) //Calculate the amount of segments to show
+                / outerRadialBarSegmentCount //Convert to normalized progress
+                * 100f; //Convert to percent (the units used by the radial bar)
             
             //Second radial bar will be treated as if it is continuous
             innerRadialBarProgress = timer2.Progress * 100f;