+++ CHANGELOG.md
@@ -32,7 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed issue with logging issue in the TimerBootstrapper in which the type of the cached currentPlayerLoop was being requested, which is always null
 
 ---
-## [0.1.0] - 2026-03-02
+## [0.1.2] - 2026-03-02
 
 ### This is a small refactoring and addition of callback events.
 
@@ -42,5 +42,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Pause** and **Resume** guarded by start-state flag (not executed if timer is not started)
 - All callbacks initially assigned to null rather than an anonymous delegate for performance purposes
 - All methods made virtual for maximum extensibility
+
+---
+## [0.1.3] - 2026-03-02
+
+### A small refactoring and addition of callback events.
+
+### Added
+- **OnTimerFinish** callback added to Timer and all its inheritors, called within a newly added Stop method that accepts the boolean finished parameter
+- Added **protected void Stop(bool finished)** method to perform all timer-stopping logic, and invoke the appropriate callback based on the finished parameter (OnTimerFinish for finished = true, OnTimerStop otherwise)
+
+### Changed
+- Adjust public **Stop()** method to simply call **Stop(false)**, retaining its original logic and callback invocation
+- In the CountdownTimer class, call Stop(true) rather than Stop() within Update() method to ensure invocation of the OnTimerFinish callback
+- Convert all delegate Action callbacks in the Timer class to Action events for safety
+- Remove/Comment out any logging/printing methods left over from development stage
 
 ---